Move EnvironmentConfiguration to own file
